# src/docs-openapi.yaml
openapi: 3.0.3
info:
  title: SMKN4 Profile API
  version: 0.1.1
  description: REST API for SMKN4 profile content management.
servers:
  - url: http://api.smkn4.sch.id
    description: Main API
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
  schemas:
    Article:
      type: object
      properties:
        message: { type: string, nullable: true }
        articles_id: { type: string, format: uuid }
        title: { type: string }
        content: { type: string }
        image_url: { type: string, nullable: true }
        published_date: { type: string, format: date-time, nullable: true }
        slug: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/ArticleStatus" }
        admin:
          {
            type: object,
            properties: { username: { type: string } },
            nullable: true,
          }
        category:
          {
            type: object,
            properties: { name: { type: string } },
            nullable: true,
          }
    ArticleStatus:
      type: string
      enum: [DRAFT, PUBLISHED]
...
    Achievement:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        publishDate: { type: string, format: date-time }
        description: { type: string }
        content: { type: string }
        image: { type: string, nullable: true }
    Announcement:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        content: { type: string }
        date: { type: string, format: date-time }
        image_url: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/ArticleStatus" }
...
    Pagination:
      type: object
      properties:
        total: { type: integer }
        totalPages: { type: integer }
        currentPage: { type: integer }
        perPage: { type: integer }
    PaginatedArticles:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Article" }
        pagination: { $ref: "#/components/schemas/Pagination" }
...
paths:
...
  /articles/:
    get:
      summary: List/search articles
      parameters:
        - in: query
          name: title
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: Paginated articles
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedArticles" }
    post:
      summary: Create article
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title, content]
              properties:
                title: { type: string }
                content: { type: string }
                category_id: { type: string, format: uuid }
                status: { $ref: "#/components/schemas/ArticleStatus" }
                image: { type: string, format: binary }
      responses:
        "200":
          {
            description: Created,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Article" } },
              },
          }
  /articles/{id}:
    put:
      summary: Update article
      security: [{ cookieAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title: { type: string }
                content: { type: string }
                category_id: { type: string, format: uuid }
                image: { type: string, format: binary }
      responses:
        "200": { description: Updated }
    delete:
      summary: Delete article
      security: [{ cookieAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Deleted }
...
  /achievements/:
    get:
      summary: List achievements with pagination or get single by ID
      parameters:
        - in: query
          name: id
          schema: { type: string, format: uuid }
          description: Get specific achievement by ID
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: Paginated achievements or single achievement
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Achievement"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Achievement" }
                      total: { type: integer }
                      page: { type: integer }
                      pageSize: { type: integer }
                      hasMore: { type: boolean }
    post:
      summary: Create achievement
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title, description, content]
              properties:
                title: { type: string }
                publishDate: { type: string, format: date-time }
                description: { type: string }
                content: { type: string, description: HTML content }
                image: { type: string, format: binary }
      responses:
        "200":
          {
            description: Created,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Achievement" } },
              },
          }
  /achievements/{id}:
    put:
      summary: Update achievement
      security: [{ cookieAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title, description, content]
              properties:
                title: { type: string }
                publishDate: { type: string, format: date-time }
                description: { type: string }
                content: { type: string }
                image: { type: string, format: binary }
      responses:
        "200": { description: Updated }
    delete:
      summary: Delete achievement
      security: [{ cookieAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Deleted }
  /announcement/:
    get:
      summary: List announcements with pagination
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200": { description: Paginated announcements }
    post:
      summary: Create announcement
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title, content]
              properties:
                title: { type: string }
                content: { type: string }
                date: { type: string, format: date-time }
                status: { $ref: "#/components/schemas/ArticleStatus" }
                image: { type: string, format: binary }
      responses:
        "200":
          {
            description: Created,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Announcement" } },
              },
          }
  /announcement/{id}:
    put:
      summary: Update announcement
      security: [{ cookieAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title: { type: string }
                content: { type: string }
                date: { type: string, format: date-time }
                status: { $ref: "#/components/schemas/ArticleStatus" }
                image: { type: string, format: binary }
      responses:
        "200": { description: Updated }
    delete:
      summary: Delete announcement
      security: [{ cookieAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Deleted }
...
security: []