# src/routes/achievements/deleteAchievement.ts
import { Router, Response } from "express";
import { prisma } from "../../config/database/prisma";
import { AuthRequest } from "../../types/auth";
import { checkAccessWithCookie } from "../../middleware/authMiddleware";
import { sendData, sendError } from "../../utils/send";
import { handlePrismaNotFound } from "../../utils/handleNotFound";
import { deleteImage } from "../../utils/imageServiceHandler";
import { BadRequestError } from "../../errorHandler/responseError";

const router = Router();

router.delete(
  "/",
  checkAccessWithCookie,
  async (req: AuthRequest, res: Response) => {
    const id = req.query.id;

    try {
      if (!id) {
        throw new BadRequestError("Invalid id");
      }

      const imageUrlToDelete = await handlePrismaNotFound(() =>
        prisma.achievements.findUnique({
          where: { id },
          select: { image_url: true },
        })
      );

      if (imageUrlToDelete.image_url) {
        deleteImage(imageUrlToDelete.image_url);
      }

      const deleted = await handlePrismaNotFound(() =>
        prisma.achievements.delete({
          where: { id },
        })
      );

      sendData(res);
    } catch (error) {
      sendError(res, error);
    }
  }
);

export default router;